+++ src/main/java/com/estacoes/estacoes/repositories/ItemRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ItemRepository extends JpaRepository<ItemContrato, Integer> {
+public interface ItemRepository extends JpaRepository<ItemContrato, Long> {
 }
